U.S. News and World Report's College Data

Statistics for a large number of US Colleges from the 1995 issue of US News and World Report.

College

Format

A data frame with 777 observations on the following 18 variables.

Private

A factor with levels No and Yes indicating private or public university

Apps

Number of applications received

Accept

Number of applications accepted

Enroll

Number of new students enrolled

Top10perc

Pct. new students from top 10% of H.S. class

Top25perc

Pct. new students from top 25% of H.S. class

F.Undergrad

Number of fulltime undergraduates

P.Undergrad

Number of parttime undergraduates

Outstate

Out-of-state tuition

Room.Board

Room and board costs

Books

Estimated book costs

Personal

Estimated personal spending

PhD

Pct. of faculty with Ph.D.'s

Terminal

Pct. of faculty with terminal degree

S.F.Ratio

Student/faculty ratio

perc.alumni

Pct. alumni who donate

Expend

Instructional expenditure per student

Grad.Rate

Graduation rate

Source

This dataset was taken from the StatLib library which is maintained at Carnegie Mellon University. The dataset was used in the ASA Statistical Graphics Section's 1995 Data Analysis Exposition.

References

James, G., Witten, D., Hastie, T., and Tibshirani, R. (2013) An Introduction to Statistical Learning with applications in R, www.StatLearning.com, Springer-Verlag, New York

Examples

summary(College)

#>  Private        Apps           Accept          Enroll       Top10perc    
#>  No :212   Min.   :   81   Min.   :   72   Min.   :  35   Min.   : 1.00  
#>  Yes:565   1st Qu.:  776   1st Qu.:  604   1st Qu.: 242   1st Qu.:15.00  
#>            Median : 1558   Median : 1110   Median : 434   Median :23.00  
#>            Mean   : 3002   Mean   : 2019   Mean   : 780   Mean   :27.56  
#>            3rd Qu.: 3624   3rd Qu.: 2424   3rd Qu.: 902   3rd Qu.:35.00  
#>            Max.   :48094   Max.   :26330   Max.   :6392   Max.   :96.00  
#>    Top25perc      F.Undergrad     P.Undergrad         Outstate    
#>  Min.   :  9.0   Min.   :  139   Min.   :    1.0   Min.   : 2340  
#>  1st Qu.: 41.0   1st Qu.:  992   1st Qu.:   95.0   1st Qu.: 7320  
#>  Median : 54.0   Median : 1707   Median :  353.0   Median : 9990  
#>  Mean   : 55.8   Mean   : 3700   Mean   :  855.3   Mean   :10441  
#>  3rd Qu.: 69.0   3rd Qu.: 4005   3rd Qu.:  967.0   3rd Qu.:12925  
#>  Max.   :100.0   Max.   :31643   Max.   :21836.0   Max.   :21700  
#>    Room.Board       Books           Personal         PhD        
#>  Min.   :1780   Min.   :  96.0   Min.   : 250   Min.   :  8.00  
#>  1st Qu.:3597   1st Qu.: 470.0   1st Qu.: 850   1st Qu.: 62.00  
#>  Median :4200   Median : 500.0   Median :1200   Median : 75.00  
#>  Mean   :4358   Mean   : 549.4   Mean   :1341   Mean   : 72.66  
#>  3rd Qu.:5050   3rd Qu.: 600.0   3rd Qu.:1700   3rd Qu.: 85.00  
#>  Max.   :8124   Max.   :2340.0   Max.   :6800   Max.   :103.00  
#>     Terminal       S.F.Ratio      perc.alumni        Expend     
#>  Min.   : 24.0   Min.   : 2.50   Min.   : 0.00   Min.   : 3186  
#>  1st Qu.: 71.0   1st Qu.:11.50   1st Qu.:13.00   1st Qu.: 6751  
#>  Median : 82.0   Median :13.60   Median :21.00   Median : 8377  
#>  Mean   : 79.7   Mean   :14.09   Mean   :22.74   Mean   : 9660  
#>  3rd Qu.: 92.0   3rd Qu.:16.50   3rd Qu.:31.00   3rd Qu.:10830  
#>  Max.   :100.0   Max.   :39.80   Max.   :64.00   Max.   :56233  
#>    Grad.Rate     
#>  Min.   : 10.00  
#>  1st Qu.: 53.00  
#>  Median : 65.00  
#>  Mean   : 65.46  
#>  3rd Qu.: 78.00  
#>  Max.   :118.00  

lm(Apps~Private+Accept,data=College)

#> 
#> Call:
#> lm(formula = Apps ~ Private + Accept, data = College)
#> 
#> Coefficients:
#> (Intercept)   PrivateYes       Accept  
#>    -170.372      182.631        1.505  
#>