Credit Card Balance Data

A simulated data set containing information on ten thousand customers. The aim here is to predict which customers will default on their credit card debt.

Credit

Format

A data frame with 10000 observations on the following 4 variables.

ID

Identification

Income

Income in $10,000's

Limit

Credit limit

Rating

Credit rating

Cards

Number of credit cards

Age

Age in years

Education

Number of years of education

Gender

A factor with levels Male and Female

Student

A factor with levels No and Yes indicating whether the individual was a student

Married

A factor with levels No and Yes indicating whether the individual was married

Ethnicity

A factor with levels African American, Asian, and Caucasian indicating the individual's ethnicity

Balance

Average credit card balance in $.

Source

Simulated data, with thanks to Albert Kim for pointing out that this was omitted, and supplying the data and man documentation page on Oct 19, 2017

References

James, G., Witten, D., Hastie, T., and Tibshirani, R. (2013) An Introduction to Statistical Learning with applications in R, www.StatLearning.com, Springer-Verlag, New York

Examples

summary(Credit)

#>        ID            Income           Limit           Rating     
#>  Min.   :  1.0   Min.   : 10.35   Min.   :  855   Min.   : 93.0  
#>  1st Qu.:100.8   1st Qu.: 21.01   1st Qu.: 3088   1st Qu.:247.2  
#>  Median :200.5   Median : 33.12   Median : 4622   Median :344.0  
#>  Mean   :200.5   Mean   : 45.22   Mean   : 4736   Mean   :354.9  
#>  3rd Qu.:300.2   3rd Qu.: 57.47   3rd Qu.: 5873   3rd Qu.:437.2  
#>  Max.   :400.0   Max.   :186.63   Max.   :13913   Max.   :982.0  
#>      Cards            Age          Education        Gender    Student  
#>  Min.   :1.000   Min.   :23.00   Min.   : 5.00    Male :193   No :360  
#>  1st Qu.:2.000   1st Qu.:41.75   1st Qu.:11.00   Female:207   Yes: 40  
#>  Median :3.000   Median :56.00   Median :14.00                         
#>  Mean   :2.958   Mean   :55.67   Mean   :13.45                         
#>  3rd Qu.:4.000   3rd Qu.:70.00   3rd Qu.:16.00                         
#>  Max.   :9.000   Max.   :98.00   Max.   :20.00                         
#>  Married              Ethnicity      Balance       
#>  No :155   African American: 99   Min.   :   0.00  
#>  Yes:245   Asian           :102   1st Qu.:  68.75  
#>            Caucasian       :199   Median : 459.50  
#>                                   Mean   : 520.01  
#>                                   3rd Qu.: 863.00  
#>                                   Max.   :1999.00  

lm(Balance ~ Student + Limit, data=Credit)

#> 
#> Call:
#> lm(formula = Balance ~ Student + Limit, data = Credit)
#> 
#> Coefficients:
#> (Intercept)   StudentYes        Limit  
#>    -334.730      404.404        0.172  
#>